fixed SideBar, layout and routing
